Extract PackageInspector from PackageCommand (#68)
Move InspectionResult construction pipeline into a new
PackageInspector.InspectAsync() in Inspectors/, following the
LibraryMetadataService pattern. PackageCommand becomes a thin
orchestrator — resolve inputs, call the inspector, format output.
